chores: retire misc functions that are not called anywhere from hicstuff module to _misc module
